requirements.txt
bidict==0.22.0
matplotlib==3.1.2
networkx==3.1
numpy==1.24.3
pandas==1.5.1
runstats==2.0.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Shapely==1.8.2
tensorboard==2.5.0
tensorboardX==2.5.1
torch-geometric==2.0.4
torch-sparse==0.6.17
tqdm==4.64.0
pyyaml==6.0.1
dtaidistance==2.3.10
protobuf==3.20.*
optuna==3.3.0